The "WebParser.dll" has been replaced by the new parser DLL "Parser.dll". This new parser DLL allows beside the parsing of a downloaded webside a given text as string.
The files "BankRegex.cs" and "DocumentRegex.cs" have been added.
This two files will later used to save parsing options for the parse of buy, sale or dividend documents.

The new file "Documents.xml" stores the parsing options for the parsing above.

This implementation is not finished yet.

The images of the buttons have been replaced by other images
The immages are taken from the website "https://icons8.de".

An installer script for the InnoSetup application has also been added.