Class TextAndImageColumn and TextAndImageCell added. This class allows to add an image to a datagridview cell.
DataGridViews.cs has been modified.
Column of the datagridview are no more generated via the databinding of the share list.
The columns are now created manually. The WKN number and share performance are TextAndImageColumn columns.

Pictures for the state if a share will be updatable or not have been added.